test(editor): Clean up duplicate DOM setup code in frontend tests
