+++ packages/frontend/editor-ui/src/__tests__/setup.ts
@@ -2,11 +2,30 @@ import '@testing-library/jest-dom';
 import 'fake-indexeddb/auto';
 import { configure } from '@testing-library/vue';
 import 'core-js/proposals/set-methods-v2';
+import { APP_MODALS_ELEMENT_ID } from '@/constants';
 
 // Avoid tests failing because of difference between local and GitHub actions timezone
 process.env.TZ = 'UTC';
 
 configure({ testIdAttribute: 'data-test-id' });
+
+// Create DOM containers for Element Plus components before each test
+beforeEach(() => {
+	// Create app-grid container for toasts
+	const appGrid = document.createElement('div');
+	appGrid.id = 'app-grid';
+	document.body.appendChild(appGrid);
+
+	// Create app-modals container for modals
+	const appModals = document.createElement('div');
+	appModals.id = APP_MODALS_ELEMENT_ID;
+	document.body.appendChild(appModals);
+});
+
+afterEach(() => {
+	// Clean up DOM containers
+	document.body.innerHTML = '';
+});
 
 window.ResizeObserver =
 	window.ResizeObserver ||
