+++ packages/frontend/editor-ui/src/components/ApiKeyCreateOrEditModal.test.ts
@@ -2,7 +2,7 @@ import { createComponentRenderer } from '@/__tests__/render';
 import { createTestingPinia } from '@pinia/testing';
 import { API_KEY_CREATE_OR_EDIT_MODAL_KEY } from '@/constants';
 import { STORES } from '@n8n/stores';
-import { cleanupAppModals, createAppModals, mockedStore, retry } from '@/__tests__/utils';
+import { mockedStore, retry } from '@/__tests__/utils';
 import ApiKeyEditModal from './ApiKeyCreateOrEditModal.vue';
 import { fireEvent } from '@testing-library/vue';
 
@@ -40,13 +40,11 @@ const settingsStore = mockedStore(useSettingsStore);
 
 describe('ApiKeyCreateOrEditModal', () => {
 	beforeEach(() => {
-		createAppModals();
 		apiKeysStore.availableScopes = ['user:create', 'user:list'];
 		settingsStore.settings.enterprise = createMockEnterpriseSettings({ apiKeyScopes: false });
 	});
 
 	afterEach(() => {
-		cleanupAppModals();
 		vi.clearAllMocks();
 	});
