+++ packages/frontend/editor-ui/src/components/ChangePasswordModal.test.ts
@@ -2,7 +2,6 @@ import { createTestingPinia } from '@pinia/testing';
 import ChangePasswordModal from '@/components/ChangePasswordModal.vue';
 import type { createPinia } from 'pinia';
 import { createComponentRenderer } from '@/__tests__/render';
-import { cleanupAppModals, createAppModals } from '@/__tests__/utils';
 
 const renderComponent = createComponentRenderer(ChangePasswordModal);
 
@@ -10,13 +9,10 @@ describe('ChangePasswordModal', () => {
 	let pinia: ReturnType<typeof createPinia>;
 
 	beforeEach(() => {
-		createAppModals();
 		pinia = createTestingPinia({});
 	});
 
-	afterEach(() => {
-		cleanupAppModals();
-	});
+	afterEach(() => {});
 
 	it('should render correctly', () => {
 		const wrapper = renderComponent({ pinia });
