+++ packages/frontend/editor-ui/src/components/ChatEmbedModal.test.ts
@@ -4,7 +4,6 @@ import { CHAT_EMBED_MODAL_KEY, WEBHOOK_NODE_TYPE } from '@/constants';
 import { STORES } from '@n8n/stores';
 import { createComponentRenderer } from '@/__tests__/render';
 import { waitFor } from '@testing-library/vue';
-import { cleanupAppModals, createAppModals } from '@/__tests__/utils';
 
 const renderComponent = createComponentRenderer(ChatEmbedModal, {
 	props: {
@@ -28,13 +27,9 @@ const renderComponent = createComponentRenderer(ChatEmbedModal, {
 });
 
 describe('ChatEmbedModal', () => {
-	beforeEach(() => {
-		createAppModals();
-	});
+	beforeEach(() => {});
 
-	afterEach(() => {
-		cleanupAppModals();
-	});
+	afterEach(() => {});
 	it('should render correctly', async () => {
 		const { getByTestId } = renderComponent();
