+++ packages/frontend/editor-ui/src/components/CommunityPackageInstallModal.test.ts
@@ -4,7 +4,7 @@ import { createTestingPinia } from '@pinia/testing';
 import { COMMUNITY_PACKAGE_INSTALL_MODAL_KEY } from '@/constants';
 import { STORES } from '@n8n/stores';
 import userEvent from '@testing-library/user-event';
-import { cleanupAppModals, createAppModals, retry } from '@/__tests__/utils';
+import { retry } from '@/__tests__/utils';
 
 const renderComponent = createComponentRenderer(CommunityPackageInstallModal, {
 	props: {
@@ -34,13 +34,9 @@ const renderComponent = createComponentRenderer(CommunityPackageInstallModal, {
 });
 
 describe('CommunityPackageInstallModal', () => {
-	beforeEach(() => {
-		createAppModals();
-	});
+	beforeEach(() => {});
 
-	afterEach(() => {
-		cleanupAppModals();
-	});
+	afterEach(() => {});
 	it('should disable install button until user agrees', async () => {
 		const { getByTestId } = renderComponent();
