+++ packages/frontend/editor-ui/src/components/FixedCollectionParameter.test.ts
@@ -1,5 +1,5 @@
 import { createComponentRenderer } from '@/__tests__/render';
-import { cleanupAppModals, createAppModals, SETTINGS_STORE_DEFAULT_STATE } from '@/__tests__/utils';
+import { SETTINGS_STORE_DEFAULT_STATE } from '@/__tests__/utils';
 import FixedCollectionParameter, { type Props } from '@/components/FixedCollectionParameter.vue';
 import { STORES } from '@n8n/stores';
 import { createTestingPinia } from '@pinia/testing';
@@ -56,14 +56,6 @@ describe('FixedCollectionParameter.vue', () => {
 		isReadOnly: false,
 	};
 	const renderComponent = createComponentRenderer(FixedCollectionParameter, { props });
-
-	beforeEach(() => {
-		createAppModals();
-	});
-
-	afterEach(() => {
-		cleanupAppModals();
-	});
 
 	it('renders the component', () => {
 		const { getByTestId } = renderComponent();
