+++ packages/frontend/editor-ui/src/components/ParameterInputFull.test.ts
@@ -3,7 +3,6 @@ import type { useNDVStore } from '@/stores/ndv.store';
 import { createTestingPinia } from '@pinia/testing';
 import type { useNodeTypesStore } from '@/stores/nodeTypes.store';
 import type { useSettingsStore } from '@/stores/settings.store';
-import { cleanupAppModals, createAppModals } from '@/__tests__/utils';
 import ParameterInputFull from './ParameterInputFull.vue';
 import { FROM_AI_AUTO_GENERATED_MARKER } from 'n8n-workflow';
 import { fireEvent } from '@testing-library/vue';
@@ -44,7 +43,6 @@ beforeEach(() => {
 		} as never,
 		isEnterpriseFeatureEnabled: { externalSecrets: false } as never,
 	};
-	createAppModals();
 });
 
 vi.mock('@/stores/ndv.store', () => {
@@ -81,11 +79,6 @@ const renderComponent = createComponentRenderer(ParameterInputFull, {
 describe('ParameterInputFull.vue', () => {
 	beforeEach(() => {
 		vi.clearAllMocks();
-		createAppModals();
-	});
-
-	afterEach(() => {
-		cleanupAppModals();
 	});
 
 	it('should render basic parameter', async () => {
