+++ packages/frontend/editor-ui/src/components/ParameterInputWrapper.test.ts
@@ -2,7 +2,7 @@ import { renderComponent } from '@/__tests__/render';
 import { createTestingPinia } from '@pinia/testing';
 import ParameterInputWrapper from './ParameterInputWrapper.vue';
 import { STORES } from '@n8n/stores';
-import { cleanupAppModals, createAppModals, SETTINGS_STORE_DEFAULT_STATE } from '@/__tests__/utils';
+import { SETTINGS_STORE_DEFAULT_STATE } from '@/__tests__/utils';
 import { waitFor } from '@testing-library/vue';
 
 vi.mock('@/composables/useWorkflowHelpers', () => {
@@ -10,14 +10,6 @@ vi.mock('@/composables/useWorkflowHelpers', () => {
 });
 
 describe('ParameterInputWrapper.vue', () => {
-	beforeEach(() => {
-		createAppModals();
-	});
-
-	afterEach(() => {
-		cleanupAppModals();
-	});
-
 	test('should resolve expression', async () => {
 		const { getByTestId } = renderComponent(ParameterInputWrapper, {
 			pinia: createTestingPinia({
