+++ packages/frontend/editor-ui/src/components/ResourceMapper.test.ts
@@ -5,7 +5,7 @@ import {
 } from './__tests__/utils/ResourceMapper.utils';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
 import type { MockedStore } from '@/__tests__/utils';
-import { cleanupAppModals, createAppModals, mockedStore, waitAllPromises } from '@/__tests__/utils';
+import { mockedStore, waitAllPromises } from '@/__tests__/utils';
 import ResourceMapper from '@/components/ResourceMapper/ResourceMapper.vue';
 import userEvent from '@testing-library/user-event';
 import { createComponentRenderer } from '@/__tests__/render';
@@ -28,14 +28,12 @@ describe('ResourceMapper.vue', () => {
 	});
 
 	beforeEach(() => {
-		createAppModals();
 		projectsStore = mockedStore(useProjectsStore);
 		projectsStore.currentProjectId = 'aProjectId';
 	});
 
 	afterEach(() => {
 		vi.clearAllMocks();
-		cleanupAppModals();
 	});
 
 	it('renders default configuration properly', async () => {
